Inject jquery automatically for pages which don't already include it
