+++ package.json
@@ -16,6 +16,7 @@
   },
   "dependencies": {
     "asyncblock": "2.2.9",
+    "jquery": "2.2.4",
     "protractor": "5.1.2",
     "webdriver-manager": "12.0.6"
   },
@@ -34,7 +35,6 @@
     "in-publish": "2.0.0",
     "jasmine-reporters": "2.1.1",
     "jasmine-spec-reporter": "2.4.0",
-    "jquery": "2.2.4",
     "load-grunt-tasks": "3.5.2",
     "mkdirp": "0.3.0",
     "tslint": "5.2.0",
